% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camazotz.R
\docType{package}
\name{camazotz-package}
\alias{camazotz}
\alias{camazotz-package}
\title{Human-readble R Objects}
\description{
\R's \code{\link{print}} and \code{\link{format}} functions can at times be
less than useful in understanding an object. \code{camazotz}, based on work
with the \code{\link[tibble]{tbl_df}} class in the \pkg{tibble} package,
faciliates the creation of human-readable outputs of \R objects.
}

