% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.R
\name{composition.data.frame}
\alias{composition.data.frame}
\title{Readable Data Frames (tibbles)}
\usage{
\method{composition}{data.frame}(x, ..., n = NULL, width = NULL, n_extra)
}
\description{
For now this function is a wrapper around \code{\link{print.tbl_df}} in
\pkg{tibble}.
}
\details{
For more information about tibbles, please refer to \code{\link{tibble}}
}

