% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{composition.numeric}
\alias{composition.integer}
\alias{composition.numeric}
\title{Readable Numeric Vectors}
\usage{
\method{composition}{numeric}(x, n = NULL, ...)

\method{composition}{integer}(x, n = NULL, ...)
}
\description{
Readable Numeric Vectors
}

