% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable.R
\name{readable}
\alias{readable}
\title{Reable R Objects}
\usage{
readable(x, ...)
}
\arguments{
\item{x}{An \R object}
}
\description{
Associate a human-readable version with a \R object. Under the hood all
\code{readable} does is add the \code{readable} to the classes of \code{x}.
}
\details{

}

